Refactor, move almost all state to backend
Previously, state was kept in localStorage in the frontend. Now, the
only (shared) state is the login token cookie. The rest is now RESTful,
meaning one can freely log in and out from different devices and still
get the correct state.

It is now also impossible to join the same game twice. Instead, one can
rejoin an active game.